ml/requirements.txt
python-dotenv
pandas
numpy
faker
sqlalchemy
bcrypt
psycopg2